Create README.md
US, Walmart, would like to predict the sales and demand accurately.
There are certain events and holidays which impact sales on each day.
There are sales data available for 45 stores of Walmart. The business is
facing a challenge due to unforeseen demands and runs out of stock
some times, due to the inappropriate machine learning algorithm.